+++ cmd/frontendbuild.sh
@@ -1,0 +1,8 @@
+cd /Users/shreysharma/go/src/github.com/glitchedgitz/cybernetic-ui
+npm run build
+cd /Users/shreysharma/go/src/github.com/glitchedgitz/grroxy/grx/frontend
+npm run build
+cd /Users/shreysharma/go/src/github.com/glitchedgitz/grroxy/cmd
+./install.sh
+cd /Users/shreysharma/go/src/github.com/glitchedgitz/grroxy/cmd/electron
+./build.sh all